refactor(agent): Fix imports
